+++ src/Symfony/NdtIdBundle.php
@@ -1,0 +1,8 @@
+<?php
+declare(strict_types=1);
+
+namespace ndtan\Symfony;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+final class NdtIdBundle extends Bundle {}